watchers:Compare annotations before discarding CiliumNode updates.
[ upstream commit 76eca7897e6e8f5edc4fb36c0a56062dc5cc4414 ]

Currently we discard CiliumNode updates based on DeepEqual and labels. However DeepEqual is set to ignore Annotations, and the wg-pub-key annotation is used to exchange rotated Wireguard keys.

The wireguard tunnel is broken when any node restart happens. After restart, the restarted node's public key got refreshed but not propagated to other nodes.

The issue is caused by cilium dropping CiliumNode update events when the spec, status and labels between the old and new nodes are the same. Wireguard public key updates are transmitted through annotations.

Signed-off-by: Lin Dong <lindongchn@gmail.com>
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>